refactor(build/tests): add tests, move build scripts to TS
