+++ package.json
@@ -12,8 +12,8 @@
   },
   "license": "MIT",
   "author": "Rob Eisenberg <rob@bluespire.com> (http://robeisenberg.com/)",
-  "main": "dist/commonjs/aurelia-router.js",
-  "module": "dist/es2015/aurelia-router.js",
+  "main": "dist/commonjs/index.js",
+  "module": "dist/es2015/index.js",
   "typings": "dist/index.d.ts",
   "repository": {
     "type": "git",
@@ -21,8 +21,9 @@
   },
   "scripts": {
     "start": "npm run dev -- --format es2015",
-    "dev": "node build/scripts/dev",
-    "build": "node build/scripts/build",
+    "dev": "npm run build -- --dev",
+    "prebuild": "rimraf dist",
+    "build": "cross-env TS_NODE_PROJECT=build/tsconfig.json ts-node build/build.ts",
     "bundle-dts": "dts-bundle-generator -o dist/index.d.ts src/index.ts",
     "karma": "karma start",
     "test": "npm run karma -- --unit --integration --single-run --browsers ChromeHeadless",
@@ -67,8 +68,10 @@
   },
   "devDependencies": {
     "@types/jasmine": "^2.8.8",
+    "@types/rimraf": "^2.0.2",
     "@types/webpack": "^4.4.17",
     "@types/webpack-env": "^1.13.6",
+    "@types/yargs": "^12.0.1",
     "aurelia-binding": "^2.1.5",
     "aurelia-bootstrapper": "^2.3.0",
     "aurelia-framework": "^1.3.0",
@@ -84,6 +87,7 @@
     "aurelia-testing": "^1.0.0",
     "aurelia-tools": "0.2.4",
     "aurelia-webpack-plugin": "^3.0.0",
+    "cross-env": "^5.2.0",
     "dts-bundle-generator": "^1.6.1",
     "gulp-watch": "^5.0.1",
     "istanbul-instrumenter-loader": "^3.0.1",
@@ -99,6 +103,7 @@
     "rollup": "^0.64.1",
     "rollup-plugin-typescript2": "^0.16.1",
     "ts-loader": "^4.4.2",
+    "ts-node": "^7.0.1",
     "tslib": "^1.9.3",
     "tslint": "^5.11.0",
     "typescript": "^3.0.1",
